requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
holidays==0.76
Pillow==10.4.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
twilio==9.6.5

# Production dependencies
mysqlclient==2.2.0
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
